Teach zerolog how to return errors, with and without stack traces.
When `zerolog` logs an error message with an `Error()` finalizer, it
returns a `error`.  If the `StackTrace()` event has been called, the
`error` will be a wrapped error using `github.com/pkg/errors` and will
include the stack trace.

There are a few subtle differences introduced with this change:

1. It is possible to defeat this behavior using the Fields() method.
2. Only one error can be handled by an Event at a time.

Time was spent considering introducing an accumulator for errors that
would be evaluated at finalizer-time, but this seemed unnecessary given
errors should be wrapped and multiple calls to `Err()` seemed unlikely
in practice.

The behavior of the `Msg()` finalizer was left in tact in order to avoid any
performance penalty in non-error handling conditions.  The deferral of error
handling can best be observed by looking at the performance hit of an `Err()`
Event with and without `StackTrace()`:

```
BenchmarkErrorContextFields-8                  	20000000	        84.8 ns/op
BenchmarkErrorContextFields_WithStackTrace-8   	  500000	      2552 ns/op
BenchmarkLogContextFields-8                    	30000000	        47.6 ns/op
PASS
ok  	github.com/rs/zerolog	4.642s
```

Fixes: #9