Add the concept of category to Event's
Introduce a new type Category (uint8) and a new `Category` method to
Event. If Event.Category is not called, the added feature has no cost
for the end-user.

However if `Event.Category` is called and if the log level assigned to
this category allows the event to be logged, this event will be logged
with an extra key `category` (configurable via CategoryFieldName),
with as value, the category name, if prealably set by client code.

Add, in `category.go` some functions to manage the category-based log
levels and category names.

SetGlobalLevel impacts category levels

There are a maximum of MaxUint8 categories and
a `Category` is a uint8...

globals: use CategoryFieldName as key to log categories