Refactored `zerolog.ConsoleWriter` to allow customization
In order to allow customizing the console output, the `zerolog.ConsoleWriter` output
has been refactored, with the formatting logic extracted to separate functions in `consoleFormatters`,
and with the `SetFormatter()` method to re-define the defaults.

Also, the default has been changed to be more sparse and visually attractive.

The logic has been extracted from an external prototype package at <https://github.com/karmi/consolelog>.

A comprehensive example has been added to the README, and a number of tests has been added as well.

Closes #84