Add support for logger-local error handlers
There are some use cases for writing to zerolog from within the
ErrorHandler, but today this is extremely tricky to do safely as it can
easily lead to ErrorHandlers triggering ErrorHandlers (causing a log
bomb). Logger-local error handlers allow us to configure eg. a primary
logger with an ErrorHandler that calls a secondary logger with no
ErrorHandler.