sampling: Support message-level sampling
zerolog's Sampler interface is currently limited
to sampling on the log level.
This makes it difficult to build a smarter sampler
that suppresses repetitive log messages
while still surfacing unique messages.
For comparison, Zap uses `(level, msg)` as the sampling key[^1],
which allows it to do the above.

[^1]: https://github.com/uber-go/zap/blob/v1.24.0/zapcore/sampler.go#L221

This change proposes a new interface, MessageSampler:

    type MessageSampler {
        Sampler

        SampleMessage(Level, string) bool
    }

If a Sampler provided to Logger implements this interface,
the SampleMessage method will be used by `*Event`
when it finally has access to the message content.

---

**Looking for opinions**:

This turns the sampler interface into a two-stage thing.
We log only if `Sample(level) && SampleMessage(level, msg) == true`.
There's risk of confusion where a SampleMessage implementation
calls Sample again--double counting that message.
Is this an acceptable risk?

Another way to do this would be to separate the two concepts
by not embedding Sampler in the MessageSampler interface.
You'd still have two points where a message can be rejected,
but because MessageSampler no longer implements Sampler,
these will be treated as independent concepts.
This will lower the risk of calling Sample from SampleMessage,
in exchange for a slightly larger API.

Resolves #543